Better layout for pyBMC information panel
Better organized pyBMC's info in a container. Populated System panel
with power switch state and power ok signal.

Changed power switches to buttons, so I can trigger a modal to confirm
operation. Added modal dialog for pyBMC's power button to select
operation (shutdown or reboot) and confirm it.

Now allowing PSU output pins to be read. That way we know if we turned
on the PSU. Note that if the PSU was turned out external to pyBMC, the
information will not be correct due to that information not being
available in the current hardware version.

Better exposing PSU information from Sensors class and from the API. Now
allowing reads from PSU output pins.

Small fix to the regex for the uptime command.