Add various CPU states in place of the throttle flag
The "vcgencmd get_throttle" command return several other flags that are
now being exposed. They are all arranged in a grid and each has an icon
to reflect the current status.

Each flag can be active (displayed with the "danger" Bootstrap style),
was active in the past (displayed with "warning" style) or never happened
("success" style).