Fix glitch when changing fan range control
For some reason, when the fan speed range text changes from 100% (the
initial value) to a new value < 100%, that causes the containing div
(.fan-speed-box) to get slightly smaller, which causes contents to shift
ever so slightly, which is very annoying. I wanted to fix this behavior
by setting the min-width of the fan-speed-box div to whatever its width
value is at load time, which is the case of when "100%" is displayed.
I couldn't get it to work though -- I tried setting min-width to
min-content, max-content, fit-content(100%), nothing seemed to work.
I was left with JavaScript as my only option, which seems to work fine.