Use WebSocket to send power switch toggle requests
Added a new message to allow for toggling PSU power through the websocket
channel. For that to work, support for sending optional command arguments
through the channel was added.

A bug in the new socket message handling exposed a bug where we were not
respecting the value of the auto-refresh toggle button when we were
about to re-create the websocket, so I fixed that too.