Allow fan control through websocket channel
Added new websocket messages to control the speed of one or more fans.

Added support for syncing fan speeds via the new "Sync Speeds" switch in
the UI. Control one and the others will follow. Only works in the UI
though (fan duty cycle set through the API/websocket won't be synced).
If the "sync speeds" switch is turned off and back on, all fans are reset
to the first fan's speed.

Also fixed a bug in the fan speed setting when the page is refreshed.
The fans' speed were being read but not updated in the UI.

Other changes:

- Ensured 0-based fan indexing throughout the project.
- Updating casing of some functions in main.js for consistency.
- Updated fan array in the payload ("fans") to "caseFans" to match the
  internal naming.
- Fixed path to fans (plural: /api/v1/fans) to adhere to the standards.