Fix power ok and fan RPM on external power switching
If power was not turned on through pyBMC, we were continuously resetting
fan state thinking that we had just lost power. Also, we were not
correctly reporting the "power Ok" state.

Well, it's possible to switch power on externally, e.g.: by pressing the
power switch on the case. In those cases, we still want to report fan
RPM and "power Ok" state. Now we do, and we also report power switch as
"On" on externally switched power and indicate as such.